move get_env_variable() from utils.py to env_var.py to avoid circular import error
