+++ newzealidar/utils.py
@@ -4,7 +4,6 @@ This module contains utility functions for the package.
 """
 import json
 import logging
-import os
 import pathlib
 import shutil
 from collections import OrderedDict
@@ -21,14 +20,14 @@ from rioxarray import merge
 # import xarray as xr
 from dotenv import load_dotenv
 import shapely
-from shapely import unary_union, to_geojson
+from shapely import unary_union
 from shapely.geometry import MultiPolygon, Polygon, GeometryCollection, box
 from sqlalchemy import create_engine
 from sqlalchemy.engine import Engine
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.pool import NullPool
 
-from newzealidar import tables
+from newzealidar import tables, env_var
 
 logger = logging.getLogger(__name__)
 
@@ -38,60 +37,9 @@ Base = declarative_base()
 
 # Generic type, used for static type checking
 Ttable = TypeVar("Ttable", bound=Base)
-T = TypeVar("T", str, bool, int, float)
 
 CATCHMENT_RESOLUTION = 30  # the resolution of the catchment in meters
 EPS = 0.1  # epsilon for float comparison
-
-
-def get_env_variable(
-    var_name: str, default: T = None, allow_empty: bool = False, cast_to: T = str
-) -> T:
-    """
-    Reads an environment variable, with settings to allow defaults, empty values, and type casting
-    To read a boolean EXAMPLE_ENV_VAR=False use get_env_variable("EXAMPLE_ENV_VAR", cast_to=bool)
-
-    :param var_name: The name of the environment variable to retrieve.
-    :param default: Default return value if the environment variable does not exist. Doesn't override empty string vars.
-    :param allow_empty: If False then a KeyError will be raised if the environment variable is empty.
-    :param cast_to: The type to cast to eg. str, int, or bool
-    :return: The environment variable, or default if it does not exist, as type T.
-    :raises: KeyError if allow_empty is False and the environment variable is empty string or None
-    :raises: ValueError if cast_to is not compatible with the value stored.
-    """
-    env_var = os.getenv(var_name, default)
-    if not allow_empty and env_var in (None, ""):
-        raise KeyError(
-            f"Environment variable {var_name} not set, and allow_empty is False"
-        )
-    return _cast_str(env_var, cast_to)
-
-
-def _cast_str(str_to_cast: str, cast_to: T) -> T:
-    """
-    Takes a string and casts it to necessary primitive builtin types. Tested with int, float, and bool.
-    For bools, this detects if the value is in the case-insensitive sets {"True", "T", "1"} or {"False", "F", "0"}
-    and raises a ValueError if not. For example _cast_str("False", bool) -> False
-
-    :param str_to_cast: The string that is going to be casted to the type
-    :param cast_to: The type to cast to e.g. bool
-    :return: The string casted to type T defined by cast_to.
-    :raises: ValueError if [cast_to] is not compatible with the value stored.
-    """
-    # Special cases i.e. casts that aren't of the form int("7") -> 7
-    if cast_to == bool:
-        # For bool we have the problem where bool("False") == True but we want this function to return False
-        truth_values = {"true", "t", "1"}
-        false_values = {"false", "f", "0"}
-        if str_to_cast.lower() in truth_values:
-            return True
-        elif str_to_cast.lower() in false_values:
-            return False
-        raise ValueError(
-            f"{str_to_cast} being casted to bool but is not in {truth_values} or {false_values}"
-        )
-    # General case
-    return cast_to(str_to_cast)
 
 
 def get_database(
@@ -124,7 +72,7 @@ def get_connection_from_profile(
         "POSTGRES_PASSWORD",
     ]
     host, port, db, username, password = (
-        get_env_variable(key) for key in connection_keys
+        env_var.get_env_variable(key) for key in connection_keys
     )
     assert (
         any(
@@ -947,8 +895,8 @@ def clip_dem(
         assert index.isdigit(), f"Catchment index {index} is not digit."
     # get DEM file path
     clipped_dem_path = (
-        pathlib.Path(get_env_variable("DATA_DIR"))
-        / pathlib.Path(get_env_variable("DEM_DIR"))
+        pathlib.Path(env_var.get_env_variable("DATA_DIR"))
+        / pathlib.Path(env_var.get_env_variable("DEM_DIR"))
         / pathlib.Path(f"{index}")
     )
     clipped_dem_path.mkdir(parents=True, exist_ok=True)
@@ -986,7 +934,7 @@ def save_gpkg(gdf: gpd.GeoDataFrame, file: Union[Type[Ttable], str]):
     """
     Save source catchments to GPKG
     """
-    gpkg_path = pathlib.Path(get_env_variable("DATA_DIR")) / pathlib.Path("gpkg")
+    gpkg_path = pathlib.Path(env_var.get_env_variable("DATA_DIR")) / pathlib.Path("gpkg")
     if isinstance(file, str):
         file_name = f"{file}.gpkg"
     else:
