+++ tests/CatchmentsTests.py
@@ -3,7 +3,7 @@ import os
 import unittest
 from unittest import mock, TestCase
 
-from src import catchments, utils
+from newzealidar import env_var, catchments
 from . import Base
 
 
@@ -14,7 +14,7 @@ class CatchmentsTests(Base, TestCase):
         """override the parameters in base class if needed."""
 
     @mock.patch.dict(os.environ, {'DATA_DIR': r'tests/data',
-                                  'POSTGRES_PORT': utils.get_env_variable('POSTGRES_PORT_TEST')})
+                                  'POSTGRES_PORT': env_var.get_env_variable('POSTGRES_PORT_TEST')})
     def test_catchments_with_json(self):
         """
         basic test of catchments module.
