+++ tests/LidarTests.py
@@ -3,7 +3,7 @@ import os
 import unittest
 from unittest import mock, TestCase
 
-from src import lidar, utils
+from newzealidar import env_var, lidar
 from . import Base
 
 
@@ -11,7 +11,7 @@ class LidarTests(Base, TestCase):
     """Tests the lidar module."""
 
     @mock.patch.dict(os.environ, {'DATA_DIR': r'tests/data',
-                                  'POSTGRES_PORT': utils.get_env_variable('POSTGRES_PORT_TEST')})
+                                  'POSTGRES_PORT': env_var.get_env_variable('POSTGRES_PORT_TEST')})
     def test_lidar(self):
         """
         basic test of lidar module.
