+++ tests/ProcessTests.py
@@ -3,7 +3,7 @@ import os
 import unittest
 from unittest import mock, TestCase
 
-from src import process, utils
+from newzealidar import env_var, process
 from . import Base
 
 
@@ -11,7 +11,7 @@ class ProcessTests(Base, TestCase):
     """Tests the process module."""
 
     @mock.patch.dict(os.environ, {'DATA_DIR': r'tests/data',
-                                  'POSTGRES_PORT': utils.get_env_variable('POSTGRES_PORT_TEST')})
+                                  'POSTGRES_PORT': env_var.get_env_variable('POSTGRES_PORT_TEST')})
     def test_process(self):
         """
         basic test of process module.
